pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "local-tts"
version = "0.1.0"
description = "Local text-to-speech API powered by Chatterbox"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "chatterbox-tts",
  "fastapi>=0.115.0",
  "setuptools<81",
  "uvicorn[standard]>=0.30.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
